+++ docs/history.md
@@ -3,6 +3,7 @@
 * Refactorized breederJob
 * Refactorized farmerJob
 * Added auto restart if bot disconnects
+* Added docker
 ## 1.2.0
 * Added Sorter Job
   * This job search nearest chest and they sort them automatically